Move common types to the model package object.
