+++ src/main/scala/net/ripe/rpki/publicationserver/parsing/PublicationMessageParser.scala
@@ -3,7 +3,7 @@ package net.ripe.rpki.publicationserver.parsing
 import java.net.URI
 
 import net.ripe.rpki.publicationserver.Binaries.{Base64, Bytes}
-import net.ripe.rpki.publicationserver._
+import net.ripe.rpki.publicationserver.model._
 
 import scala.annotation.tailrec
 import scala.io.Source
